% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeds.R
\docType{data}
\name{feeds}
\alias{feeds}
\title{List of RSS feeds.}
\format{A data frame with 72 rows and 1 variable}
\usage{
feeds
}
\description{
A dataset containing a list of feeds, both RSS and Atom, and a few
included that are not real feeds and so produce an error.
}
\details{
\itemize{
  \item feeds \code{character}. Feed urls.
}
}
\keyword{datasets}
